Update CSES Sum of Divisors.mdx
I have updated the editorial for the problem Sum of Divisors after trying the problem myself and getting stuck on it. In my case I found it necessary to seek for external help via USACO forum / other editorials to fully understand the problem. Hence I proposed an update to the editorial.