Update USACO_FAQs.mdx
Resolved all suggestions, I am applying to be on the website development team and its my first time doing something this big as a sophomore in hs. Sorry for making the faq's a bit vauge, I can assure you i will addapt to the way we write things like this faq.